test(ui): consolidate the providers page integration suites
- Merge the GCP and hierarchy suites into the providers page file,
  moving every describe verbatim
- Hoist the shared fixtures, helpers and imports to file scope
- Establish one integration file per real page
